requirements_.txt
absl-py==1.3.0
altair==4.2.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1660605382950/work
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1660619017351/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.6
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
etils==0.9.0
exceptiongroup==1.0.4
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1661768596365/work
flatbuffers==22.11.23
fonttools==4.38.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-images-download @ git+http://github.com/joeclinton1/google-images-download.git@e91e6a38ad654877f59edc7894822f5319d75df1
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1661804640198/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1662481517711/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1661522530937/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1658332349812/work
keras==2.11.0
keras-cv==0.3.4
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1662384672173/work
protobuf==3.19.6
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1662356143277/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
Pympler==1.0.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyrsistent==0.19.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1660328968266/work
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
selenium==4.6.0
semver==2.13.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1661624565828/work
streamlit==1.15.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.12.0
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.7.1
torchvision==0.8.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1656937829196/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1655411388954/work
trio==0.22.0
trio-websocket==0.9.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.11.0
